Some data added

+++ server/schema/sections.schema.ts
@@ -1,10 +1,10 @@
 import { relations } from 'drizzle-orm';
-import { integer, pgTable, smallserial, varchar } from 'drizzle-orm/pg-core';
+import { integer, pgTable, serial, varchar } from 'drizzle-orm/pg-core';
 
 import { faculty, staff } from '.';
 
 export const sections = pgTable('sections', {
-  id: smallserial('id').primaryKey(),
+  id: serial('id').primaryKey(),
   name: varchar('name', { length: 100 }).notNull(),
   aboutUs: varchar('about_us').notNull(),
   headFacultyId: integer('head_faculty_id')
